+++ kernel/src/mem_function.h
@@ -6,7 +6,7 @@
 
 void *memcpy(void *dest, const void *src, size_t n);
 
-void *memset(void *s, int c, size_t n);
+void *memset(void *s, uint8_t c, size_t n);
 
 void *memmove(void *dest, const void *src, size_t n);
